+++ inc/vapor/gui/Label.h
@@ -44,7 +44,7 @@ public:
 	void setText( const std::string& text );
 
 	// Updates the geometry to match the new text if needed.
-	virtual void update( float delta );
+	virtual void update( double delta );
 
 	// Serializes this node to a stream.
 	virtual const std::string save( int indent = 0 );
