+++ inc/vapor/scene/Transform.h
@@ -84,7 +84,7 @@ public:
 	bool requiresBoundingVolumeUpdate() const;
 
 	// Called once per frame to update the component.
-	virtual void update( float delta );
+	virtual void update( double delta );
 
 	// Gets the type of this component. 
 	virtual const std::string& getType() const;
