+++ src/Engine/Console.cpp
@@ -45,7 +45,7 @@ void Console::buildGeometry()
 
 //-----------------------------------//
 
-void Console::update( float UNUSED(delta) )
+void Console::update( double UNUSED(delta) )
 {
 	// No need to update geometry if the label did not change.
 	if( !isDirty ) return;
